Refactor knowledge agent and remove deprecated kg_maintainer module
- Move key mapping and normalization logic from kg_maintainer to knowledge_agent
- Remove unused kg_maintainer package and its modules (merge, models, parsing)
- Update imports in knowledge_agent to reflect new structure
- Add missing utils import for normalize_trait_name function
- Consolidate character and world update key mappings in knowledge_agent
