--- kg_maintainer/__init__.py
@@ -1,33 +1,0 @@
-"""Package consolidating KG maintainer utilities."""
-
-from .merge import (
-    merge_character_profile_updates,
-    merge_world_item_updates,
-)
-from .models import (
-    AgentStateData,
-    CharacterProfile,
-    EvaluationResult,
-    PatchInstruction,
-    ProblemDetail,
-    SceneDetail,
-    WorldItem,
-)
-from .parsing import (
-    parse_unified_character_updates,
-    parse_unified_world_updates,
-)
-
-__all__ = [
-    "AgentStateData",
-    "CharacterProfile",
-    "EvaluationResult",
-    "PatchInstruction",
-    "ProblemDetail",
-    "SceneDetail",
-    "WorldItem",
-    "parse_unified_character_updates",
-    "parse_unified_world_updates",
-    "merge_character_profile_updates",
-    "merge_world_item_updates",
-]
